[nrf noup] Adjust range for builtin keys
Opens the full range of vendor keys to be used via builtin key
mechanism.

This change makes the version of the Oberon PSA core 1.3.4 incompatible
with the volatile keys when MBEDTLS_PSA_STATIC_KEY_SLOTS is NOT enabled.

Since we decided to set the starting range of the builtin keys to
be the first section of the vendor key id range (PSA_KEY_ID_VENDOR_*)
these now overlaped with the volatile keys. So to fix this we
move the volatile key ids after the builtin key ids.

Signed-off-by: Vidar Lillebø <vidar.lillebo@nordicsemi.no>
Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>