[nrf toup] Update key type for the key_wrapping functions
The newly added key wrapping functions oberon_psa_wrap_key
and oberon_psa_unwrap_key used the psa_key_id_t type for the
key objects instead of using the mbedtls_svc_key_id_t type which
is used in all the other functions.

This breaks TF-M builds.

This is communicated to Oberon and they agreed that the current patch
will be included in the next version of the Oberon PSA core.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>