[nrf noup] library: Add _Static_assert in common.h
GCC supports _Static_assert from version 4.6 as a builtin feature
even when you compile with std=99.

Use this _Static_assert for the implementation of
MBEDTLS_STATIC_ASSERT if there is no defined static_assert.

I opened a PR in the MbedTLS PSA core to include that:
https://github.com/Mbed-TLS/TF-PSA-Crypto/pull/190

Hopefully they will accept that and then we can remove
this commit in the next release.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>