[nrf noup] Add static_asserts for key slot size
Adds a static assert to ensure that the config
MBEDTLS_PSA_STATIC_KEY_SLOT_BUFFER_SIZE is large enough
to fit the enabled key types.

This is a noup for now but we are in discussion with Oberon
in order to integrate this to the core in the next release.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>