[nrf toup] fix unused variable warning
Fix an unused variable warning that happens when building with the LLVM
toolchain.
Adding the `unused` attribute to the variable to mark it as potentially
unused is the simplest as it's used in different `#ifdef` scenarios.

This should be communicated to Oberon and fixed upstream.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>
(cherry picked from commit a258e58246da0b34c7cb6fa27358807ffcd079ed)
Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>