[nrf noup] include: Remove driver contexts files
Remove the driver contexts from the Oberon repo so that
we don't rely on the include folder order for these files.

This is a noup for now but I contacted Oberon and they said
that they will remove these headers from here and move these
files in a different directory so this should be a short
lived noup.

Ref: NCSDK-33148

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>