[nrf toup] don't try to find volatile/builtin keys from wrong sources
When not finding a given key in memory, the implementation would try
to find it from the persistent keys regardless of the actual key type
(volatile/builtin/persistent).

Don't try to find inexistent volatile/builtin keys from persistent
ones.

In addition to the calls being superflous, the problem that was
happening here is that the ITS implementation (Secure Storage
subsystem) returns `PSA_ERROR_INVALID_ARGUMENT` because the ID is not
in the persistent key range, and because it doesn't return
`PSA_ERROR_DOES_NOT_EXIST` then the wrong error code is propagated
back to the caller.

`toup` as the issue has been communicated to Mbed TLS and should be
fixed there.
See:
- https://github.com/Mbed-TLS/TF-PSA-Crypto/issues/488
- https://github.com/Mbed-TLS/TF-PSA-Crypto/pull/492

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>