[nrf noup] Add built-in key support to psa_copy_key
If the target key ID is in the built-in range, then creation of the key
should be handled by a driver. Otherwise, there's no way for the key to
be persisted.

I don't know if this is acceptable upstream, since Mbed TLS and Oberon
openly state that they don't support creating built-in keys through the
PSA Crypto API, even though there are ways around this limitation.

At first, out of abundance of caution, this change will only be applied
for a single platform: Haltium (nRF54H, nRF92).

Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>